Skip two_hop in consolidation mode

+++ sky/jobs/controller.py
@@ -1870,32 +1870,23 @@ class ControllerManager:
                 # we continue to try cleaning up whatever else we can.
 
             # Clean up any files mounted from the local disk, such as two-hop
-            # file mounts.
-            for file_mount in (task.file_mounts or {}).values():
-                try:
-                    # Skip if we are using cloud storage as the source.
-                    if data_utils.is_cloud_store_url(file_mount):
-                        continue
-                    # Otherwise, we always cleanup local files since they are
-                    # no longer needed after task cleanup, the file can be:
-                    # - Two hop file mounts rsynced from the API server: refer
-                    #   translate_local_file_mounts_to_two_hop for more details.
-                    # - API server file mount cache in consolidation mode:
-                    #   actually there is a dummy two hop that rsync the files
-                    #   from ~/.sky/clients to ~/.sky/tmp/controller/{ID}
-                    #   on server, which isolates the file mounts between
-                    #   tasks. Here we assume the source is always isolated
-                    #   even if the dummy two hop is removed.
-                    # TODO(aylei): remove dummy two hop after we isolate the
-                    # file mount cache for tasks.
-                    path = os.path.expanduser(file_mount)
-                    if os.path.isdir(path):
-                        shutil.rmtree(path)
-                    else:
-                        os.remove(path)
-                except Exception as e:  # pylint: disable=broad-except
-                    logger.warning(
-                        f'Failed to clean up file mount {file_mount}: {e}')
+            # file mounts for non-consolidatoin mode.
+            # For consolidation mode, the file_mounts are shared across
+            # workloads and the lifecycle will be managed by API server.
+            if not managed_job_utils.is_consolidation_mode():
+                for file_mount in (task.file_mounts or {}).values():
+                    try:
+                        # Skip if we are using cloud storage as the source.
+                        if data_utils.is_cloud_store_url(file_mount):
+                            continue
+                        path = os.path.expanduser(file_mount)
+                        if os.path.isdir(path):
+                            shutil.rmtree(path)
+                        else:
+                            os.remove(path)
+                    except Exception as e:  # pylint: disable=broad-except
+                        logger.warning(
+                            f'Failed to clean up file mount {file_mount}: {e}')
 
             if error is not None:
                 raise error
